// src/hansei-intelligence-processor/raindrop.gen.ts
// Code generated by raindrop build, DO NOT EDIT.
import { Ai, Annotation, App, KvCache, Logger, MRNObject, SmartMemory, SmartSql, Tracer,  } from '@liquidmetal-ai/raindrop-framework';

export interface Env {
  _raindrop: {
    app: App;
  };
  AGENT_MEMORY: SmartMemory;
  AI: Ai;
  annotation: Annotation<Omit<MRNObject, 'type' | 'applicationName' | 'versionId'>>;
  GRAPH_DATABASE: SmartSql;
  logger: Logger;
  mem: KvCache;
  SESSION_CACHE: KvCache;
  tracer: Tracer;
}